+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change log
 This is the changelog for [AWS Architect](readme.md).
 
+## 5.0 ##
+* Upgrade to OpenAPI-factory 3.0, see [openapi factory](https://github.com/wparad/openapi-factory.js/blob/master/CHANGELOG.md#30) for breaking changes.  This means that the authorizer now takes `event` and the authorization token must be manually extracted.
+* Support authorizer resolution in express `server.run`
+
 ## 4.2 ##
 * Include cloud formation deployment for standard resources.
 * Added example cloud formation template to template service.